Changed grade to double

+++ src/main/java/com/app/myschool/model/SubjectView.java
@@ -40,7 +40,7 @@ public class SubjectView {
 
 	private Integer qtrGradeType;
 
-	private Integer qtrGrade;
+	private Double qtrGrade;
 
 	private Boolean qtrLocked;
